--- js/components/product-list.js
@@ -1,126 +1,0 @@
-//import data from './data.js';
-
-//let items = data.items;
-
-// if(!items) {
-//     data.items = createItems();
-//     items = data.items;
-// }
-
-function createItems() {
-    let items = [{
-        id: 1,
-        name: 'bag',
-        image: '/images/bag.jpg',
-        viewed: 0,
-        votes: 0
-    }, {
-        id: 2,
-        name: 'banana',
-        image: 'images/banana.jpg',
-        viewed: 0,
-        votes: 0,
-    }, {
-        id: 3,
-        name: 'bathroom',
-        image: 'images/bathroom.jpg',
-        viewed: 0,
-        votes: 0,
-    }, {
-        id: 4,
-        name: 'boots',
-        image: 'images/boots.jpg',
-        viewed: 0,
-        votes: 0
-    }, {
-        id: 5,
-        name: 'breakfast',
-        image: 'images/breakfast.jpg',
-        viewed: 0,
-        votes: 0
-    }, {
-        id: 6,
-        name: 'bubblegum',
-        image: '/images/bubblegum.jpg',
-        viewed: 0,
-        votes: 0
-    }, {
-        id: 7,
-        name: 'chair',
-        image: '/images/cthulhu.jpg',
-        viewed: 0,
-        votes: 0
-    }, {
-        id: 8,
-        name: 'dog-duck',
-        image: '/images/dog-duck.jpg',
-        viewed: 0,
-        votes: 0
-    }, {
-        id: 9,
-        name: 'dragon',
-        image: 'images/dragon.jpg',
-        viewed: 0,
-        votes: 0
-    }, {
-        id: 10,
-        name: 'pen',
-        image: '/images/pen.jpg',
-        viewed: 0,
-        votes: 0
-    }, { 
-        id: 11,
-        name: 'pet-sweep',
-        image:'/images/pet-sweep.jpg',
-        viewed: 0,
-        votes: 0
-    }, {
-        id: 12,
-        name: 'scissors',
-        image: '/images/scissors.jpg',
-        viewed: 0,
-        votes:  0
-    }, {
-        id: 13,
-        name: 'shark',
-        image: '/images/shark.jpg',
-        viewed: 0,
-        votes: 0
-    }, {
-        id: 14,
-        name: 'sweep',
-        image: '/images/sweep.jpg',
-        viewed: 0,
-        votes: 0
-    }, {
-        id: 15,
-        name: 'tauntaun',
-        image: '/images/tauntaun.jpg',
-        viewed: 0,
-        votes: 0
-    }, {
-        id: 16,
-        name: 'unicorn',
-        image: '/images/unicorn.jpg',
-        viewed: 0,
-        votes: 0
-    }, {
-        id: 17,
-        name: 'usb',
-        image: '/images/usb.gif',
-        viewed: 0,
-        votes: 0
-    }, {
-        id: 18,
-        name: 'water-can',
-        image: 'images/water-can.jpg',
-        viewed: 0,
-        votes:  0
-    }, {
-        id: 19,
-        name: 'wine-glass',
-        image: 'images/wine-glass.jpg',
-        viewed: 0,
-        votes: 0
-    }];
-}
